+++ pkg/controller/operatingsystemconfig/generator/ignition/generator.go
