+++ pkg/controller/operatingsystemconfig/generator/ignition/generator_test.go
