+++ pkg/controller/operatingsystemconfig/generator/templates/cloud-init.sh.template
